// src/common/enums/error-codes.enum.ts
export enum ErrorCode {
  /* PRISMA ERROR */
  PRISMA_EMAIL_CREATE_ERROR = 'MS_EMAIL_PRISMA_CREATE_EMAIL_ERROR',
  PRISMA_EMAIL_UPDATE_ERROR = 'MS_EMAIL_PRISMA_UPDATE_EMAIL_ERROR',
  PRISMA_EMAIL_DELETE_ERROR = 'MS_EMAIL_PRISMA_DELETE_EMAIL_ERROR',

  /* TEMPLATE ERROR */
  TEMPLATE_NOT_FOUND = 'MS_EMAIL_TEMPLATE_NOT_FOUND',
  TEMPLATE_CANNOT_ACCESS = 'MS_EMAIL_TEMPLATE_CANNOT_ACCESS',
  TEMPLATE_UNKNOWN_ERROR = 'MS_EMAIL_TEMPLATE_UNKNOWN_ERROR',
  TEMPLATE_PATH_INVALID = 'MS_EMAIL_TEMPLATE_PATH_INVALID',

  /* NESTJS MAILER */
  NESTJSMAILER_SENDING_FAILED = 'MS_EMAIL_NESTJSMAILER_SENDING_FAILED',
  NESTJSMAILER_AUTHENTICATION_FAILED = 'MS_EMAIL_NESTJSMAILER_AUTHENTICATION_FAILED',

  /* VALIDATION */
  INVALID_PAYLOAD = 'MS_EMAIL_INVALID_PAYLOAD',
}